pyproject.toml
[project]
name = "backend"
version = "2.5.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.63",
    "caseutil>=0.7.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "wagtail>=7.1",
    "wagtail-localize>=1.12.2",
]

[dependency-groups]
dev = [
    "django-cors-headers>=4.7.0",
    "invoke>=2.2.0",
    "ruff>=0.11.7",
]

[tool.ruff]
exclude = [".venv", ".git", ".vscode", "__pycache__"]
line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
extend-select = ["I"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
